+++ eis-jianming-master/src/main/java/com/prolog/eis/service/impl/unbound/calculation/CalculationService.java
@@ -3,9 +3,10 @@ package com.prolog.eis.service.impl.unbound.calculation;
 import com.prolog.eis.dto.sxk.OutBoundSxStoreDto;
 import com.prolog.eis.service.impl.unbound.entity.DetailDataBean;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public interface CalculationService<T> {
-    public List<T> buildOutBoundQty(DetailDataBean detailDataBeand, float miniPackage, boolean isPickStation);
+    public List<T> buildOutBoundQty(DetailDataBean detailDataBeand, BigDecimal miniPackage, boolean isPickStation);
 
 }
